+++ packages/trading-hours/contracts/interfaces/ITradingHoursPlugin.sol
@@ -3,20 +3,21 @@ pragma solidity >=0.5.0;
 
 /// @title ITradingHoursPlugin
 /// @notice Public interface for the Trading Hours plugin module
-/// @dev Blocks swap, add liquidity and flash outside configured trading hours.
+/// @dev Blocks swap outside configured trading hours. Flash and liquidity operations (add and remove)
+/// are always allowed.
 /// While disabled (see setEnabled), trading is fully unrestricted, including the weekend rule - the
 /// connector checks this flag directly and skips the verification call entirely when disabled.
 /// Once enabled: trading hours and blocked windows are plain UTC seconds-of-day. Weekends (Sat/Sun) are
 /// always blocked and not configurable - the weekend offset only shifts which UTC day counts as local Sat/Sun.
 /// Additional per-day blocked windows (holidays, temporary closures) can be set for specific UTC days.
-/// Remove liquidity is always allowed.
 interface ITradingHoursPlugin {
-  /// @param day Any UTC timestamp within the target day - floored to the start of that day
+  /// @param day Any UTC timestamp within the target day - floored to the start of that day. uint32 is
+  /// enough for any UTC timestamp before year 2106
   /// @param index Slot index, 0 to 4
   /// @param startSeconds UTC seconds from midnight when the window starts, inclusive. (0, 0) clears the slot
   /// @param endSeconds UTC seconds from midnight when the window ends, exclusive
   struct BlockedWindowInput {
-    uint256 day;
+    uint32 day;
     uint8 index;
     uint24 startSeconds;
     uint24 endSeconds;
@@ -33,7 +34,7 @@ interface ITradingHoursPlugin {
 
   /// @notice Emitted when a blocked window for a specific day is set or cleared
   /// @param day UTC timestamp of the start of the day this window applies to
-  event BlockedWindowUpdated(uint256 indexed day, uint8 indexed index, uint24 startSeconds, uint24 endSeconds);
+  event BlockedWindowUpdated(uint32 indexed day, uint8 indexed index, uint24 startSeconds, uint24 endSeconds);
 
   /// @notice The current time falls outside the configured trading window
   error TradingNotAllowed();
@@ -67,7 +68,7 @@ interface ITradingHoursPlugin {
   /// @param index Slot index, 0 to 4
   /// @param startSeconds UTC seconds from midnight when the window starts, inclusive. (0, 0) clears the slot
   /// @param endSeconds UTC seconds from midnight when the window ends, exclusive
-  function setBlockedWindow(uint256 day, uint8 index, uint24 startSeconds, uint24 endSeconds) external;
+  function setBlockedWindow(uint32 day, uint8 index, uint24 startSeconds, uint24 endSeconds) external;
 
   /// @notice Set or clear multiple blocked windows, across any mix of days and slots, in one transaction
   function setBlockedWindows(BlockedWindowInput[] calldata inputs) external;
@@ -84,7 +85,7 @@ interface ITradingHoursPlugin {
   /// @notice Returns one of the blocked windows for a specific day
   /// @param day Any UTC timestamp within the target day - floored to the start of that day
   /// @param index Slot index, 0 to 4
-  function getBlockedWindow(uint256 day, uint8 index) external view returns (uint24 startSeconds, uint24 endSeconds);
+  function getBlockedWindow(uint32 day, uint8 index) external view returns (uint24 startSeconds, uint24 endSeconds);
 
   /// @notice Returns whether trading is allowed at the given timestamp
   /// @param timestamp UTC timestamp to check
